Detect Player on Sound
Added a code that detects the player's new location and compares its distance to the "hearing range" of Ceratos. Also changed the viewAngle to 180 for more "human" like sight (e.g. if i was in the center and he was walking around the outside he should realistically see me) so that it works like a "peripheral" sight.